fix(round20-7): negative unpacked-array low bound silently dropped a word
`int a[-1:1]` held two elements, not three. `foreach` ran twice and reported
nothing — exit 0, no diagnostic. This was filed as a LOUD gap because touching
`a[-1]` explicitly did emit an E4002; the pure-`foreach` form is what showed it
was silent (iverilog 3, vita 2).

Root cause: `array_dim_extents` clamped the folded low bound through
`clamp_bound_u32`, so `[-1:1]` became `(lo 0, size 2)` — the clamp shrank the
DIM, not just the base. `lo` is `i64` now through `array_dims`, `net_dim_extents`,
both `flatten_word` twins and the `net_dims` sidecar, so the declared index
space is spoken end to end: access, `foreach`, `$size`/`$left`/`$right`/`$low`/
`$high`, and per-element VCD names (`a[-1]`, not a renumbered `a[0]`).

Byte identity for everything that already worked: the per-dim guard and
coordinate moved into two shared helpers whose `lo >= 0` branches are the old
code verbatim, in the old push order (`lo == 0` still emits no `Sub`). Only a
negative-`lo` dim takes the new path — coordinate first, then a `coord <u size`
guard, which is exact whether or not the user's index expression is signed.
The routed string-array form declined a negative bound *because* `flatten_word`
carried `u32`; that decline is lifted, so `string s[-1:1]` gets runtime index +
`foreach` too.

format_version 23 → 24: `NetDimsTable` is a `.velab` trailer and postcard
encodes `u32` as a plain varint but `i64` zigzag, so a stale artifact would
mis-decode rather than fail. sim-ir UNCHANGED (schema hash / canonical / RON
goldens untouched) — same class as the v20-v23 trailer changes.

Two pre-existing defects surfaced by making the tuple signed, both fixed:
- `$bits` of a partially-indexed view read `array_dims`'s `(lo, SIZE)` as
  `(lo, hi)`: `int a[2][3]` → `$bits(a[0])` = 128, iverilog 96.
- the engine's file-I/O base did the same, as `lo.min(size)`. It cancels out
  when no address is given, so only the ranged form showed it: `$readmemh` into
  `reg m[10:11]` with `10, 11` warned and left `xx xx` where iverilog loads
  `aa bb`. A NEGATIVE base there is now a warn-and-bail — those addresses are
  `u64` end to end — rather than a silent read from word 0.

And one self-inflicted: dropping the negative-bound decline exposed an i64
overflow it had been masking (`[4611686018427387904:-4611686018427387905]`),
caught by the test that exists for exactly that. Checked arithmetic.

Adversarial review, 2 lenses:
- differential: 16249 scratchpad designs PRE vs POST — 20 changed (7 from the
  fork fix, 13 here), every one now matching iverilog. Zero regressions.
- soundness: 1-D/multi-dim, ascending/descending, const/runtime index, string
  and int element types, array query functions, VCD naming, `$readmem`. A
  genuine OOB (`a[2]` on `[-1:1]`) stays loud + X, as iverilog warns + x.

4565 tests green (9 new) · clippy 0 · fmt 0.

+++ crates/elaborate/src/arrays.rs
@@ -575,7 +575,7 @@ impl Elaborator<'_> {
     pub(crate) fn flatten_assign_pattern<'a>(
         &mut self,
         elems: &'a [ast::Expr],
-        dims: &[(u32, u32)],
+        dims: &[(i64, u32)],
     ) -> Option<Vec<&'a ast::Expr>> {
         let size = dims[0].1 as usize;
         if elems.len() != size {
